unlabeled tags and daily stock

+++ app/src/main/java/com/loyalstring/modelclasses/Item.java
@@ -13,10 +13,10 @@ public class Item {
     private String status;
     @SerializedName("GrossWeight")
     @Expose
-    private Integer grossWeight;
+    private double grossWeight;
     @SerializedName("NetWeight")
     @Expose
-    private Integer netWeight;
+    private double netWeight;
     @SerializedName("Quantity")
     @Expose
     private Integer quantity;
@@ -79,19 +79,19 @@ public class Item {
         this.status = status;
     }
 
-    public Integer getGrossWeight() {
+    public double getGrossWeight() {
         return grossWeight;
     }
 
-    public void setGrossWeight(Integer grossWeight) {
+    public void setGrossWeight(double grossWeight) {
         this.grossWeight = grossWeight;
     }
 
-    public Integer getNetWeight() {
+    public Double getNetWeight() {
         return netWeight;
     }
 
-    public void setNetWeight(Integer netWeight) {
+    public void setNetWeight(double netWeight) {
         this.netWeight = netWeight;
     }
